+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/api/WorktableAPI.java
@@ -1,6 +1,7 @@
 package com.codetaylor.mc.artisanworktables.modules.worktables.api;
 
 import com.codetaylor.mc.artisanworktables.modules.worktables.recipe.RegistryRecipeWorktable;
+import com.codetaylor.mc.artisanworktables.modules.worktables.reference.EnumWorktableType;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -16,11 +17,18 @@ public class WorktableAPI {
     for (EnumWorktableType type : values) {
       RECIPE_REGISTRY_MAP.put(type.getName(), new RegistryRecipeWorktable());
     }
+
+    RECIPE_REGISTRY_MAP.put("mage", new RegistryRecipeWorktable());
   }
 
   public static RegistryRecipeWorktable getRecipeRegistry(EnumWorktableType type) {
 
-    return RECIPE_REGISTRY_MAP.get(type.getName());
+    return WorktableAPI.getRecipeRegistry(type.getName());
+  }
+
+  public static RegistryRecipeWorktable getRecipeRegistry(String name) {
+
+    return RECIPE_REGISTRY_MAP.get(name);
   }
 
   private WorktableAPI() {