+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.7.10:
+* Added: new tables: engineers worktable, mages worktable
+* Added: new tools: engineers spanner, engineers driver, mages athame, mages grimoire
+* Changed: lightened the wood tool color slightly
+* Note: don't forget to regenerate the config or manually add the new tool types to the list
+
 1.6.10:
 * Fixed: extra chance drops are now properly linked in JEI; when searching for recipes to make the extra chance drop, it will show the worktable recipe (#14)
 * Fixed: double-click same item as craft result will not cause a craft to happen (#15)